requirements.txt

flask>=3.1.0
matplotlib>=3.10.1
yfinance>=0.2.55
pandas>=2.0.3
python-dateutil>=2.8.2
seaborn
pytest>=7.0.0